+++ mythtv/libs/libmythtv/channelscan/scaninfo.cpp
@@ -32,7 +32,7 @@ uint SaveScan(const ScanDTVTransportList &scan)
 
     // Delete very old scans
     const vector<ScanInfo> list = LoadScanList();
-    for (uint i = 0; i < list.size(); ++i)
+    for (size_t i = 0; i < list.size(); ++i)
     {
         if (list[i].m_scandate > MythDate::current().addDays(-14))
             continue;
@@ -63,7 +63,7 @@ uint SaveScan(const ScanDTVTransportList &scan)
     if (!scanid)
         return scanid;
 
-    for (uint i = 0; i < scan.size(); ++i)
+    for (size_t i = 0; i < scan.size(); ++i)
         scan[i].SaveScan(scanid);
 
     return scanid;
