+++ mythtv/libs/libmythtv/frequencytables.cpp
@@ -259,7 +259,7 @@ freq_table_list_t get_matching_freq_tables(
         get_matching_freq_tables_internal(format, modulation, country);
 
     freq_table_list_t new_list;
-    for (uint i = 0; i < list.size(); i++)
+    for (size_t i = 0; i < list.size(); i++)
         new_list.push_back(new FrequencyTable(*list[i]));
 
     return new_list;
@@ -274,7 +274,7 @@ long long get_center_frequency(
     freq_table_list_t list =
         get_matching_freq_tables_internal(format, modulation, country);
 
-    for (uint i = 0; i < list.size(); ++i)
+    for (size_t i = 0; i < list.size(); ++i)
     {
         int min_freqid = list[i]->m_nameOffset;
         int max_freqid = min_freqid +
@@ -296,7 +296,7 @@ int get_closest_freqid(
     freq_table_list_t list =
         get_matching_freq_tables_internal(format, modulation, country);
 
-    for (uint i = 0; i < list.size(); ++i)
+    for (size_t i = 0; i < list.size(); ++i)
     {
         int min_freqid = list[i]->m_nameOffset;
         int max_freqid = min_freqid +
